feature #139 : 로그인, 로그아웃 구현
